Creating, alter and delete querys, entities, repositories, services, resources and DTOs

+++ src/main/java/com/devdavicosta/teaminfoapi/services/StadiumService.java
@@ -4,10 +4,16 @@ import java.util.List;
 import java.util.Optional;
 
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.dao.DataIntegrityViolationException;
 import org.springframework.stereotype.Service;
 
 import com.devdavicosta.teaminfoapi.entities.Stadium;
+import com.devdavicosta.teaminfoapi.entities.State;
 import com.devdavicosta.teaminfoapi.repositories.StadiumRepository;
+import com.devdavicosta.teaminfoapi.services.exceptions.DatabaseException;
+import com.devdavicosta.teaminfoapi.services.exceptions.ResourceNotFoundException;
+
+import jakarta.persistence.EntityNotFoundException;
 
 @Service
 public class StadiumService {
@@ -15,13 +21,24 @@ public class StadiumService {
 	@Autowired
 	private StadiumRepository repository;
 	
+	@Autowired
+	private StateService stateService;
+	
 	public List<Stadium> findAll() {
 		return repository.findAll();
 	}
 	
 	public Stadium findById(Long id) {
 		Optional<Stadium> obj = repository.findById(id);
-		return obj.get();
+		return obj.orElseThrow(() -> new ResourceNotFoundException(id));
+	}
+	
+	public List<Stadium> findByName(String text) {
+		return repository.findByStadiumName(text);
+	}
+	
+	public List<Stadium> findByState(String text) {
+		return repository.findByState(text);
 	}
 	
 	public Stadium insert(Stadium obj) {
@@ -29,18 +46,28 @@ public class StadiumService {
 	}
 	
 	public Stadium update(Long id, Stadium obj) {
-		Stadium entity = repository.getReferenceById(id);
-		updateData(entity, obj);
-		return repository.save(entity);
+		try {
+			Stadium entity = repository.getReferenceById(id);
+			updateData(entity, obj);
+			return repository.save(entity);
+		} catch (EntityNotFoundException e) {
+			throw new ResourceNotFoundException(id);
+		}
 	}
 	
 	private void updateData(Stadium entity, Stadium obj) {
 		entity.setNome(obj.getNome());
 		entity.setNome_popular(obj.getNome_popular());
-		entity.setEstado(obj.getEstado());
+		State state = stateService.findById(obj.getEstado().getId());
+		entity.setEstado(state);
 	}
 	
 	public void delete(Long id) {
-		repository.deleteById(id);
+		try {
+			 if (!repository.existsById(id)) throw new ResourceNotFoundException(id);
+			 repository.deleteById(id);
+		} catch (DataIntegrityViolationException e) {
+			 throw new DatabaseException(e.getMessage());
+		}
 	}
 }
